Create test users via script
