+++ backend/shared/src/helpers/on-create-user.ts
@@ -1,9 +1,9 @@
 import { getUser, isProd } from 'shared/utils'
 import { type PrivateUser, type User } from 'common/user'
 import { createSupabaseDirectClient } from 'shared/supabase/init'
-import { createPrivateUserMessageChannelMain } from 'api/create-private-user-message-channel'
+import { createPrivateUserMessageChannelMain } from '@manifold/api/lib/create-private-user-message-channel'
 import { type JSONContent } from '@tiptap/core'
-import { createPrivateUserMessageMain } from 'api/create-private-user-message'
+import { createPrivateUserMessageMain } from '@manifold/api/lib/create-private-user-message'
 import { addNewUserToLeague } from 'shared/generate-leagues'
 import { createReferralsProgramNotification } from 'shared/create-notification'
 import { sendWelcomeEmail } from 'shared/emails'
